move player history endpoint to /players/me/history and off AuthRoutes
